+++ src/ncvue/ncvscatter.py
@@ -48,8 +48,12 @@ from .ncvwidgets import add_spinbox, add_tooltip
 import matplotlib
 # matplotlib.use('TkAgg')
 from matplotlib import pyplot as plt
-# plt.style.use('seaborn-darkgrid')
-plt.style.use('seaborn-dark')
+try:
+    # plt.style.use('seaborn-v0_8-darkgrid')
+    plt.style.use('seaborn-v0_8-dark')
+except OSError:
+    # plt.style.use('seaborn-darkgrid')
+    plt.style.use('seaborn-dark')
 # plt.style.use('fast')